+++ lib/models/commands_calls_models/configure_command_model.dart
@@ -7,6 +7,7 @@ class ConfigureCommandModel {
   bool autoUpdate = false;
   bool isDebug = false;
   bool skipFirebaseConfigure = false;
+  bool skipShorebirdConfigure = false;
   bool skipPubUpdate = false;
   bool skipVersionUpdate = false;
 
@@ -18,6 +19,10 @@ class ConfigureCommandModel {
     isDebug = argResults?[ClonifyCommandFlags.isDebug.name] as bool? ?? false;
     skipFirebaseConfigure =
         argResults?[ClonifyCommandFlags.skipFirebaseConfigure.name] as bool? ??
+        false;
+    skipShorebirdConfigure =
+        argResults?[ClonifyCommandFlags.skipShorebirdConfigure.name]
+            as bool? ??
         false;
     skipPubUpdate =
         argResults?[ClonifyCommandFlags.skipPubUpdate.name] as bool? ?? false;